ChatBoat/requirements.txt
flask
flask-cors
tensorflow
numpy
pillow
kagglehub
googletrans==4.0.0-rc1
werkzeug
